+++ CosmicWorks.sln
@@ -3,19 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.12.35728.132 d17.12
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "models", "models\models.csproj", "{2E0F21A9-8440-45A1-8C26-4AD996207B00}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "modeling-demos", "modeling-demos\modeling-demos.csproj", "{08A0EE80-3546-44EC-B771-105C6CC5D0A7}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9EDC2DA7-DD5F-4D6D-A058-C25CAC8A3964}"
-	ProjectSection(SolutionItems) = preProject
-		azuredeploy.json = azuredeploy.json
-		config.json = config.json
-		init-sandbox.sh = init-sandbox.sh
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cosmos-management", "cosmos-management\cosmos-management.csproj", "{0330BF2A-28BB-4FBC-8226-B742DDA3F455}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CosmicWorks", "src\CosmicWorks.csproj", "{334F18E2-2B0E-4A4E-9E03-0150882C7C62}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,23 +11,12 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2E0F21A9-8440-45A1-8C26-4AD996207B00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E0F21A9-8440-45A1-8C26-4AD996207B00}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E0F21A9-8440-45A1-8C26-4AD996207B00}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E0F21A9-8440-45A1-8C26-4AD996207B00}.Release|Any CPU.Build.0 = Release|Any CPU
-		{08A0EE80-3546-44EC-B771-105C6CC5D0A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{08A0EE80-3546-44EC-B771-105C6CC5D0A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{08A0EE80-3546-44EC-B771-105C6CC5D0A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{08A0EE80-3546-44EC-B771-105C6CC5D0A7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0330BF2A-28BB-4FBC-8226-B742DDA3F455}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0330BF2A-28BB-4FBC-8226-B742DDA3F455}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0330BF2A-28BB-4FBC-8226-B742DDA3F455}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0330BF2A-28BB-4FBC-8226-B742DDA3F455}.Release|Any CPU.Build.0 = Release|Any CPU
+		{334F18E2-2B0E-4A4E-9E03-0150882C7C62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{334F18E2-2B0E-4A4E-9E03-0150882C7C62}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{334F18E2-2B0E-4A4E-9E03-0150882C7C62}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{334F18E2-2B0E-4A4E-9E03-0150882C7C62}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {48A2E007-4CC2-4AA3-88AE-194625C328F9}
 	EndGlobalSection
 EndGlobal
