--- config.json
@@ -1,111 +1,0 @@
-{
-  "resources": {
-    "databases": [
-      {
-        "name": "cosmicworks-v1",
-        "maxThroughput": 4000,
-        "containers": [
-          {
-            "name": "customer",
-            "pk": "/id"
-          },
-          {
-            "name": "customerAddress",
-            "pk": "/id"
-          },
-          {
-            "name": "customerPassword",
-            "pk": "/id"
-          },
-          {
-            "name": "product",
-            "pk": "/id"
-          },
-          {
-            "name": "productCategory",
-            "pk": "/id"
-          },
-          {
-            "name": "productTag",
-            "pk": "/id"
-          },
-          {
-            "name": "productTags",
-            "pk": "/id"
-          },
-          {
-            "name": "salesOrder",
-            "pk": "/id"
-          },
-          {
-            "name": "salesOrderDetail",
-            "pk": "/id"
-          }
-        ]
-      },
-      {
-        "name": "cosmicworks-v2",
-        "maxThroughput": 4000,
-        "containers": [
-          {
-            "name": "customer",
-            "pk": "/id"
-          },
-          {
-            "name": "product",
-            "pk": "/categoryId"
-          },
-          {
-            "name": "productCategory",
-            "pk": "/type"
-          },
-          {
-            "name": "productTag",
-            "pk": "/type"
-          },
-          {
-            "name": "salesOrder",
-            "pk": "/customerId"
-          }
-        ]
-      },
-      {
-        "name": "cosmicworks-v3",
-        "maxThroughput": 4000,
-        "containers": [
-          {
-            "name": "customer",
-            "pk": "/customerId"
-          },
-          {
-            "name": "product",
-            "pk": "/categoryId"
-          },
-          {
-            "name": "productMeta",
-            "pk": "/type"
-          }
-        ]
-      },
-      {
-        "name": "cosmicworks-v4",
-        "maxThroughput": 4000,
-        "containers": [
-          {
-            "name": "customer",
-            "pk": "/customerId"
-          },
-          {
-            "name": "product",
-            "pk": "/categoryId"
-          },
-          {
-            "name": "productMeta",
-            "pk": "/type"
-          }
-        ]
-      }
-    ]
-
-  }
-}
