--- cosmos-management/cosmos-management.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <RootNamespace>cosmos_management</RootNamespace>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-	<ItemGroup>
-		<PackageReference Include="Azure.Identity" Version="1.12.1" />
-		<PackageReference Include="Azure.ResourceManager.CosmosDB" Version="1.4.0-beta.10" />
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.2" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.2" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.2" />
-		<PackageReference Include="Microsoft.Graph" Version="5.58.0" />
-	</ItemGroup>
-</Project>
