--- data/.devcontainer/devcontainer.json
@@ -1,17 +1,0 @@
-{
-  "image": "mcr.microsoft.com/devcontainers/dotnet:9.0",
-  "features": {
-    "ghcr.io/azure/azure-dev/azd": {},
-    "ghcr.io/devcontainers/features/azure-cli": {},
-    "ghcr.io/devcontainers/features/docker-in-docker": {}
-  },
-  "customizations": {
-    "vscode": {
-      "extensions": [
-        "ms-azuretools.azure-dev",
-        "ms-azuretools.vscode-cosmosdb",
-        "ms-dotnettools.csdevkit"
-      ]
-    }
-  }
-}
