--- modeling-demos/appSettings.json
@@ -1,8 +1,0 @@
-﻿{
-  "subscriptionId": "",
-  "resourceGroup": "",
-  "accountName": "",
-  "location": "",
-  "uri": "",
-  "key": ""
-}
