+++ src/appSettings.json
@@ -1,0 +1,7 @@
+﻿{
+  "ACCOUNT_ENDPOINT": "",
+  "SUBSCRIPTION_ID": "",
+  "RESOURCE_GROUP": "",
+  "LOCATION": "",
+  "ACCOUNT_NAME": ""
+}
